+++ CHANGELOG.md
@@ -10,6 +10,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - New linters
    - TestNameWithTooManyCharactersLinter
 
+- Linters that are impacted by `Rule` elements in Gherkin have been updated, now that Rules models are available 
+  in CukeModeler
+  - FeatureWithoutScenariosLinter
+
 ## [1.2.1] - 2021-06-13
 
 ### Fixed